move handlers to their own file, move core capabilities to toolkit
