+++ mdb_toolkit/__init__.py
@@ -6,9 +6,11 @@ logger = logging.getLogger(__name__)
 logger.info("Initializing mdb_toolkit package")
 
 from .core import CustomMongoClient, Node, Edge
+from .MultiModalRetriever import MultiModalRetriever
 
 __all__ = [
     'CustomMongoClient',
     'Node',
-    'Edge'
+    'Edge',
+    'MultiModalRetriever'
 ]
